MainViewModel refactor
