+++ WolvenKit.App/WolvenKit.App.csproj
@@ -235,6 +235,7 @@
     <Compile Include="Model\IWolvenkitView.cs" />
     <Compile Include="Model\ProcessHelper.cs" />
     <Compile Include="Model\SearchParameters.cs" />
+    <Compile Include="Model\WccHelper.cs" />
     <Compile Include="Model\WKPackage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controllers\RadishConfiguration.cs" />
